#360: extract Compress module (unzip/zipped) from god-class
